+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/client/prisonApi/model/SentenceAndOffenceDetails.kt
@@ -1,0 +1,5 @@
+package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.client.prisonApi.model
+
+class SentenceAndOffenceDetails {
+  val sentenceTypeDescription: String? = null
+}
