--- src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/response/model/SentenceAndOffenceDetails.kt
@@ -1,5 +1,0 @@
-package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.response.model
-
-class SentenceAndOffenceDetails {
-  val sentenceTypeDescription: String? = null
-}
